moving folders to natural places, creating format function for notion response
