--- src/constants.js
@@ -1,2 +1,0 @@
-export const CONTACT_ENDPOINT =
-  'https://public.herotofu.com/v1/fd08ea40-afa4-11ec-b83f-8f17e10d6288'
