+++ utils/constants.js
@@ -1,0 +1,24 @@
+export const CONTACT_ENDPOINT =
+  'https://public.herotofu.com/v1/fd08ea40-afa4-11ec-b83f-8f17e10d6288'
+
+export const GUEST_LIST_PROPERTIES = {
+  NUM_ADULTS: '# of Adults?',
+  NUM_CHILDREN: '# of Children?',
+  ASSOCIATION: 'Association',
+  CHILD_1: 'Child 1️⃣',
+  CHILD_2: 'Child 2️⃣',
+  CITY: 'City',
+  EMAIL: 'Email',
+  FIRST_NAME: 'First Name',
+  LAST_NAME: 'Last Name',
+  PARTNER_FIRST_NAME: 'Partner First Name',
+  PARTNER_LAST_NAME: 'Partner Last Name',
+  SENT_ANNOUNCEMENT: 'Sent Announcement?',
+  STATE: 'State / Region',
+  STREET_ADDRESS: 'Street Address',
+  STREET_ADDRESS_2: 'Street Address, line 2',
+  THANK_YOU: 'Thank You?',
+  TOTAL_IN_PARTY: 'Total in Party',
+  TYPE: 'Type',
+  WEBSITE_VISITS: 'Website Visits',
+}
